% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.checkEnum.R
\name{constraints.checkEnum}
\alias{constraints.checkEnum}
\title{Check enum}
\usage{
constraints.checkEnum(constraint, value)
}
\arguments{
\item{constraint}{constraint}

\item{value}{value}
}
\value{
TRUE value in constraint
}
\description{
Check enum
}
