% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.checkMaximum.R
\name{constraints.checkMaximum}
\alias{constraints.checkMaximum}
\title{Check if maximum constraint is met}
\usage{
constraints.checkMaximum(constraint, value)
}
\arguments{
\item{constraint}{constraint}

\item{value}{value}
}
\value{
TRUE if value is less than the constraint
}
\description{
Check if maximum constraint is met
}
