% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.checkMaxLength.R
\name{constraints.checkMaxLength}
\alias{constraints.checkMaxLength}
\title{Check Maximum Length of Character/s according to a constraint}
\usage{
constraints.checkMaxLength(constraint, value)
}
\value{
TRUE if the constraint of maximum character length is met
}
\description{
Check Maximum Length of Character/s according to a constraint
}
