% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.checkMinimum.R
\name{constraints.checkMinimum}
\alias{constraints.checkMinimum}
\title{Check if minimum constraint is met}
\usage{
constraints.checkMinimum(constraint, value)
}
\arguments{
\item{constraint}{constraint}

\item{value}{value}
}
\value{
TRUE if value is less than the constraint
}
\description{
Check if minimum constraint is met
}
