% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.checkMinLength.R
\name{constraints.checkMinLength}
\alias{constraints.checkMinLength}
\title{Check Minimum Length of Character/s according to a constraint}
\usage{
constraints.checkMinLength(constraint, value)
}
\arguments{
\item{constraint}{constraint}

\item{value}{value}
}
\value{
TRUE if the constraint of minimum character length is met
}
\description{
Check Minimum Length of Character/s according to a constraint
}
