% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.checkPattern.R
\name{constraints.checkPattern}
\alias{constraints.checkPattern}
\title{Check a Character pattern according to a constraint}
\usage{
constraints.checkPattern(constraint, value)
}
\value{
TRUE if the constraint of character pattern is met
}
\description{
Check a Character pattern according to a constraint
}
