% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.checkUnique.R
\name{constraints.checkUnique}
\alias{constraints.checkUnique}
\title{Check if a field is unique}
\usage{
constraints.checkUnique(constraint, value)
}
\arguments{
\item{constraint}{constraint}

\item{value}{value}
}
\value{
TRUE if field is unique
}
\description{
Check if a field is unique
}
