% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castNumber.R
\docType{data}
\name{CURRENCY_CHAR}
\alias{CURRENCY_CHAR}
\title{default currecy char}
\format{An object of class \code{character} of length 3.}
\usage{
CURRENCY_CHAR
}
\description{
default currecy char
}
\keyword{datasets}
