% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castNumber.R
\docType{data}
\name{DEFAULT_DECIMAL_CHAR}
\alias{DEFAULT_DECIMAL_CHAR}
\title{default decimal char}
\format{
An object of class \code{character} of length 1.
}
\usage{
DEFAULT_DECIMAL_CHAR
}
\description{
default decimal char
}
\keyword{datasets}
