% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castNumber.R
\docType{data}
\name{DEFAULT_GROUP_CHAR}
\alias{DEFAULT_GROUP_CHAR}
\title{default group char}
\format{
An object of class \code{character} of length 1.
}
\usage{
DEFAULT_GROUP_CHAR
}
\description{
default group char
}
\keyword{datasets}
