% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castDate.R
\docType{data}
\name{DEFAULT_PATTERN}
\alias{DEFAULT_PATTERN}
\title{default date pattern}
\format{An object of class \code{character} of length 1.}
\usage{
DEFAULT_PATTERN
}
\description{
default date pattern
}
\keyword{datasets}
