% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castDuration.R
\name{Duration}
\alias{Duration}
\title{duration}
\usage{
Duration(x, type = "seconds")
}
\arguments{
\item{x}{input duration}

\item{type}{time unit of x}
}
\description{
duration
}
\details{
Available inputs for type of input duration are picoseconds, nanoseconds, microseconds, milliseconds, seconds, minutes, hours, days, weeks, years
}
