% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castDuration.R
\name{durations}
\alias{durations}
\title{Durations}
\usage{
durations(years = 0, months = 0, days = 0, hours = 0, minutes = 0, seconds = 0)
}
\arguments{
\item{years}{years}

\item{months}{months}

\item{days}{days}

\item{hours}{hours}

\item{minutes}{minutes}

\item{seconds}{seconds}
}
\description{
Help function to use with \code{\link{types.castDuration}}
}
\seealso{
\code{\link{types.castDuration}}
}
