% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castBoolean.R
\docType{data}
\name{FALSE_VALUES}
\alias{FALSE_VALUES}
\title{false values}
\format{An object of class \code{character} of length 5.}
\usage{
FALSE_VALUES
}
\description{
false values
}
\keyword{datasets}
