% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get.field.descriptor.properties}
\alias{get.field.descriptor.properties}
\title{Extract the field descriptors properties}
\usage{
get.field.descriptor.properties(descriptor)
}
\arguments{
\item{descriptor}{descriptor}
}
\description{
Extract the field descriptors properties
}
