% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{has_name_field_descriptor}
\alias{has_name_field_descriptor}
\title{check if name property is missing}
\usage{
has_name_field_descriptor(descriptor)
}
\arguments{
\item{descriptor}{descriptor}
}
\description{
check if name property is missing
}
