% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helpers.expandFieldDescriptor}
\alias{helpers.expandFieldDescriptor}
\title{Expand Field Descriptor}
\usage{
helpers.expandFieldDescriptor(descriptor)
}
\arguments{
\item{descriptor}{descriptor}
}
\description{
Expand Field Descriptor
}
