% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helpers.expandSchemaDescriptor}
\alias{helpers.expandSchemaDescriptor}
\title{Expand Schema Descriptor}
\usage{
helpers.expandSchemaDescriptor(descriptor)
}
\arguments{
\item{descriptor}{descriptor}
}
\description{
Helper function to expand schema descriptor
}
