% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{class}
\name{Helpers}
\alias{Helpers}
\title{Helpers class}
\format{\code{\link{R6Class}} object.}
\usage{
Helpers
}
\value{
Object of \code{\link{R6Class}} .
}
\description{
Helpers class
}
\keyword{data}
