% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helpers.retrieveDescriptor}
\alias{helpers.retrieveDescriptor}
\title{Retrieve Descriptor}
\usage{
helpers.retrieveDescriptor(descriptor)
}
\arguments{
\item{descriptor}{descriptor}
}
\description{
Helper function to retrieve descriptor
}
