% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R, R/validate.R
\name{infer}
\alias{infer}
\alias{validate}
\title{infer}
\usage{
infer(source, options = list())

validate(descriptor)
}
\arguments{
\item{source}{source}

\item{options}{options}

\item{source}{source}

\item{options}{options}
}
\description{
infer

infer
}
