% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.integer.R
\docType{class}
\name{Integer}
\alias{Integer}
\title{Integer class}
\format{\code{\link{R6Class}} object.}
\usage{
Integer
}
\value{
Object of \code{\link{R6Class}} .
}
\description{
Integer class
}
\keyword{data}
