% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.maximum.R
\docType{class}
\name{Maximum}
\alias{Maximum}
\title{Maximum class}
\format{\code{\link{R6Class}} object.}
\usage{
Maximum
}
\value{
Object of \code{\link{R6Class}} .
}
\description{
Maximum class
}
\keyword{data}
