% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.minimum.R
\docType{class}
\name{Minimum}
\alias{Minimum}
\title{Minimum class}
\format{\code{\link{R6Class}} object.}
\usage{
Minimum
}
\value{
Object of \code{\link{R6Class}} .
}
\description{
Minimum class
}
\keyword{data}
