% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.number.R
\docType{class}
\name{Number}
\alias{Number}
\title{Number class}
\format{\code{\link{R6Class}} object.}
\usage{
Number
}
\value{
Object of \code{\link{R6Class}} .
}
\description{
Number class
}
\keyword{data}
