% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castNumber.R
\docType{data}
\name{PERCENT_CHAR}
\alias{PERCENT_CHAR}
\title{default percent char}
\format{An object of class \code{character} of length 6.}
\usage{
PERCENT_CHAR
}
\description{
default percent char
}
\keyword{datasets}
