% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\docType{class}
\name{Profile}
\alias{Profile}
\title{Profile class}
\format{
\code{\link{R6Class}} object.
}
\usage{
# Profile$load(profile)
}
\arguments{
\item{profile}{String: the name of a profile from the registry or a URL pointing to a JSON Schema.}
}
\value{
Object of class \code{\link{R6Class}}.
}
\description{
Class to represent a JSON Schema profile from the \href{https://specs.frictionlessdata.io/schemas/registry.json}{Profiles Registry}.
}
\section{Methods}{

\describe{
  \item{\code{Profile$new(descriptor = descriptor)}}{Use \code{Profile$load} to instantiate a \code{Profile} class.}

  \item{\code{validate(descriptor)}}{Validate a tabular data package descriptor against the \code{Profile}.
    \describe{
      \item{\code{descriptor}}{The dereferenced tabular data package descriptor to validate.}
      \item{Return value}{Returns \code{TRUE} if the descriptor is valid; otherwise returns \code{FALSE} with an error message.}
    }
  }
}
}

\section{Properties}{

\describe{
  \item{\code{name}}{Returns the profile name, if available.}
  \item{\code{jsonschema}}{Returns the contents of the profile's JSON Schema.}
}
}

\seealso{
\href{https://specs.frictionlessdata.io//profiles/}{Profile Specifications}
}
\keyword{data}
