% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readable.array.R
\docType{class}
\name{ReadableArray}
\alias{ReadableArray}
\title{ReadableArray class}
\format{\code{\link{R6Class}} object.}
\usage{
ReadableArray
}
\value{
Object of \code{\link{R6Class}} .
}
\description{
ReadableArray class
}
\keyword{data}
