% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readable.connection.R
\docType{class}
\name{ReadableConnection}
\alias{ReadableConnection}
\title{ReadableConnection class}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} .
}
\description{
Readable connection class
}
\keyword{data}
