% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.string.R
\docType{class}
\name{String}
\alias{String}
\title{String class}
\format{\code{\link{R6Class}} object.}
\usage{
String
}
\value{
Object of \code{\link{R6Class}} .
}
\description{
String class
}
\keyword{data}
