% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{Table.load}
\alias{Table.load}
\title{load table}
\usage{
Table.load(source, schema = NULL, strict = FALSE, headers = 1, ...)
}
\arguments{
\item{source}{source}

\item{schema}{schema}

\item{strict}{strict}

\item{headers}{headers}

\item{...}{other arguments to pass}
}
\description{
load table
}
