% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableschemaerror.R
\docType{class}
\name{TableSchemaError}
\alias{TableSchemaError}
\title{TableSchemaError class}
\format{
\code{\link{R6Class}} object.
}
\arguments{
\item{message}{message}

\item{error}{error}
}
\value{
Object of \code{\link{R6Class}} .
}
\description{
Error class for Table Schema
}
\section{Fields}{

\describe{
\item{\code{message}}{}

\item{\code{error}}{}
}}

\keyword{data}
