% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castBoolean.R
\docType{data}
\name{TRUE_VALUES}
\alias{TRUE_VALUES}
\title{default true values}
\format{
An object of class \code{character} of length 4.
}
\usage{
TRUE_VALUES
}
\description{
default true values
}
\keyword{datasets}
