% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castArray.R
\name{types.castArray}
\alias{types.castArray}
\title{Cast array}
\usage{
types.castArray(format, value)
}
\arguments{
\item{format}{no options (other than the default)}

\item{value}{lists, or valid JSON format arrays to cast}
}
\description{
Cast array is used for list objects
}
\seealso{
\code{\link{types.castList}}, 
\href{https://frictionlessdata.io/specs/table-schema/#array}{Types and formats specifications}
}
