% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castBoolean.R
\name{types.castBoolean}
\alias{types.castBoolean}
\title{cast boolean}
\usage{
types.castBoolean(format, value, options = { })
}
\arguments{
\item{format}{format}

\item{value}{value}

\item{options}{options specify trueValues or/and falseValues}
}
\description{
cast boolean
}
