% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castGeopoint.R
\name{types.castGeopoint}
\alias{types.castGeopoint}
\title{cast geopoint}
\usage{
types.castGeopoint(format, value)
}
\arguments{
\item{format}{format}

\item{value}{value}
}
\description{
cast geopoint
}
