% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castInteger.R
\name{types.castInteger}
\alias{types.castInteger}
\title{cast integer}
\usage{
types.castInteger(format, value, options = { })
}
\arguments{
\item{format}{format}

\item{value}{value}

\item{options}{options bareNumber}
}
\description{
cast integer
}
