% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castList.R
\name{types.castList}
\alias{types.castList}
\title{Cast list}
\usage{
types.castList(format, value)
}
\arguments{
\item{format}{no options (other than the default)}

\item{value}{lists, or valid JSON format arrays to cast}
}
\description{
cast list
}
\examples{

types.castList(format = "default", value =  list())

types.castList(format = "default", value = list('key', 'value'))

types.castList(format = "default", value = '["key", "value"]') # cast valid json array

}
\seealso{
\href{https://frictionlessdata.io/specs/table-schema/#array}{Types and formats specifications}
}
