% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castNumber.R
\name{types.castNumber}
\alias{types.castNumber}
\title{cast number}
\usage{
types.castNumber(format, value, options = { })
}
\arguments{
\item{format}{format}

\item{value}{value}

\item{options}{options}
}
\description{
cast number
}
