% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castObject.R
\name{types.castObject}
\alias{types.castObject}
\title{Cast object}
\usage{
types.castObject(format, value)
}
\arguments{
\item{format}{no options (other than the default)}

\item{value}{object to cast}
}
\description{
Cast object data which is lists or valid JSON.
}
\examples{

types.castObject(format = "default", value = list())

types.castObject(format = "default", value = "{}")

types.castObject(format = "default", value = '{"key": "value"}')

}
\seealso{
\href{https://frictionlessdata.io/specs/table-schema/#object}{Types and formats specifications}
}
