% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castTime.R
\name{types.castTime}
\alias{types.castTime}
\title{Cast time without a date}
\usage{
types.castTime(format = "\%H:\%M:\%S", value)
}
\arguments{
\item{format}{available options are "default", "any", and "<pattern>" where
\describe{
\item{\code{default }}{An ISO8601 time string e.g. hh:mm:ss}
\item{\code{any }}{As for \code{\link{types.castDate}}}
\item{\code{<pattern> }}{As for \code{\link{types.castDate}}}
}}

\item{value}{time to cast}
}
\description{
Cast time without a date
}
\examples{

types.castTime(format = "default", value = '06:00:00')

}
\seealso{
\href{https://frictionlessdata.io/specs/table-schema/#time}{Types and formats specifications},
\code{\link[base]{strptime}}, \code{\link[base]{DateTimeClasses}},
\code{\link[parsedate]{parsedate-package}} and 
\code{\link[lubridate]{lubridate-package}}.
}
