% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castYearmonth.R
\name{types.castYearmonth}
\alias{types.castYearmonth}
\title{cast yearmonth}
\usage{
types.castYearmonth(value)
}
\arguments{
\item{value}{value}

\item{format}{format}
}
\description{
cast yearmonth
}
