% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{validate descriptor}
\usage{
validate(descriptor)
}
\arguments{
\item{descriptor}{schema descriptor, one of:
\itemize{
\item string with the local CSV file (path)
\item string with the remote CSV file (url)
\item list object
}}
}
\value{
\code{TRUE} on valid
}
\description{
Validates whether a schema is a validate Table Schema accordingly to the specifications. 
It does not validate data against a schema.
}
