% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writable.R
\docType{class}
\name{Writeable}
\alias{Writeable}
\title{Writeable class}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} .
}
\description{
Writable streams class
}
\keyword{data}
