% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.checkEnum.R
\name{constraints.checkEnum}
\alias{constraints.checkEnum}
\title{Check Enum}
\usage{
constraints.checkEnum(constraint, value)
}
\arguments{
\item{constraint}{numeric list,matrix or vector with the constraint values}

\item{value}{numeric value to meet the constraint}
}
\value{
TRUE if value meets the constraint
}
\description{
Check if the value is exactly match a constraint.
}
\examples{

constraints.checkEnum(constraint = list(1, 2), value = 1)

constraints.checkEnum(constraint = list(1, 2), value = 3)
}
\seealso{
\href{https://frictionlessdata.io/specs/table-schema/#constraints}{Constraints specifications}
}
