% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.checkPattern.R
\name{constraints.checkPattern}
\alias{constraints.checkPattern}
\title{evaluate value over a constraint}
\usage{
constraints.checkPattern(constraint, value)
}
\arguments{
\item{constraint}{constraint}

\item{value}{value}
}
\value{
TRUE if the constraint of character pattern is met
}
\description{
evaluate value over a constraint
}
