% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.checkRequired.R
\name{constraints.checkRequired}
\alias{constraints.checkRequired}
\title{Check Minimum Length of Character/s according to a constraint}
\usage{
constraints.checkRequired(constraint, value)
}
\value{
TRUE if the constraint of minimum character length is met
}
\description{
Check Minimum Length of Character/s according to a constraint
}
