% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.checkUnique.R
\name{constraints.checkUnique}
\alias{constraints.checkUnique}
\title{is unique}
\usage{
constraints.checkUnique(x)
}
\arguments{
\item{constraint}{constraint}

\item{value}{value}
}
\value{
TRUE if unique
}
\description{
is unique
}
