% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.checkUnique.R
\name{constraints.checkUnique}
\alias{constraints.checkUnique}
\title{Check if a field is unique}
\usage{
constraints.checkUnique(constraint, value)
}
\arguments{
\item{constraint}{set TRUE to check unique values}

\item{value}{value to check}
}
\value{
TRUE if field is unique
}
\description{
If \code{TRUE}, then all values for that field \code{MUST} be unique within the data file in which it is found.
}
\examples{

constraints.checkUnique(constraint = FALSE, value = "any")

constraints.checkUnique(constraint = TRUE, value = "any")

}
\seealso{
\href{https://frictionlessdata.io/specs/table-schema/#constraints}{Constraints specifications}
}
