% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\docType{class}
\name{Constraints}
\alias{Constraints}
\title{Constraints class}
\format{\code{\link{R6Class}} object.}
\usage{
Constraints
}
\value{
Object of \code{\link{R6Class}} .
}
\description{
Constraints class
}
\keyword{data}
