% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readable.R
\docType{class}
\name{Readable}
\alias{Readable}
\title{Readable class}
\format{\code{\link{R6Class}} object.}
\usage{
Readable
}
\value{
Object of \code{\link{R6Class}}.
}
\description{
Readable class that allows typed access to its members
}
\keyword{data}
