% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema.load}
\alias{schema.load}
\title{load schema}
\usage{
schema.load(descriptor = "", strict = FALSE,
  caseInsensitiveHeaders = FALSE)
}
\arguments{
\item{descriptor}{descriptor}

\item{strict}{strict}

\item{caseInsensitiveHeaders}{caseInsensitiveHeaders}
}
\description{
load schema
}
