% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableschemaerror.R
\docType{class}
\name{TableSchemaError}
\alias{TableSchemaError}
\title{TableSchemaError class}
\format{\code{\link{R6Class}} object.}
\usage{
TableSchemaError
}
\value{
Object of \code{\link{R6Class}} .
}
\description{
TableSchemaError class
}
\keyword{data}
