% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castDate.R
\name{types.castDate}
\alias{types.castDate}
\title{cast date}
\usage{
types.castDate(format, value)
}
\arguments{
\item{format}{specify format, default is "\%Y-\%m-\%d"}

\item{value}{value}
}
\description{
cast date
}
