% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castDuration.R
\name{types.castDuration}
\alias{types.castDuration}
\title{cast duration}
\usage{
types.castDuration(format = "default", value)
}
\arguments{
\item{format}{time unit of x}

\item{value}{value}
}
\description{
cast duration
}
\details{
Available inputs for format of input value are picoseconds, nanoseconds, microseconds, milliseconds, seconds, minutes, hours, days, weeks, years
}
