% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.castYearmonth.R
\name{types.castYearmonth}
\alias{types.castYearmonth}
\title{cast yearmonth}
\usage{
types.castYearmonth(format = "\%y-\%m", value)
}
\arguments{
\item{format}{format}

\item{value}{value}
}
\description{
cast yearmonth
}
